Gui: Make profiler section into window class
