+++ src/Game.cpp
@@ -411,8 +411,8 @@ bool Game::Run()
 			    /*drawBoundingBoxes =*/_config.drawBoundingBoxes,
 			    /*cullBack =*/false,
 			    /*bgfxDebug =*/_config.bgfxDebug,
+			    /*bgfxProfile =*/_config.bgfxProfile,
 			    /*wireframe =*/_config.wireframe,
-			    /*profile =*/_config.showProfiler,
 			    /*timeOfDay =*/_config.timeOfDay,
 			    /*bumpMapStrength =*/_config.bumpMapStrength,
 			    /*smallBumpMapStrength =*/_config.smallBumpMapStrength,
