+++ src/Game.h
@@ -112,7 +112,6 @@ public:
 
 		bool wireframe {false};
 		bool waterDebug {false};
-		bool showProfiler {false};
 		bool showLandIsland {false};
 		bool showVillagerNames {false};
 		bool debugVillagerNames {false};
@@ -132,6 +131,7 @@ public:
 		float smallBumpMapStrength {1.0f};
 
 		bool bgfxDebug {false};
+		bool bgfxProfile {false};
 		bool running {false};
 
 		uint32_t numFramesToSimulate {0};
