+++ src/Gui/DebugWindow.h
@@ -28,6 +28,7 @@ class DebugWindow
 {
 public:
 	DebugWindow(std::string name, ImVec2 defaultSize);
+	DebugWindow(const DebugWindow&) = delete;
 	virtual ~DebugWindow();
 
 	void WindowUpdate(Game& game, const Renderer& renderer);
